* remotes/origin/dev/robust:
  add super_res_robust_function.cxx
  changes from patch set 4
  add variable beta
  remove lines giving warnings
  modify config files
  remove comments
  add MEDIANDATA_IMAGEPRIOR
  resize frames_median
  color shift with median
  add frame_step
  upsample ref image
  intensity shift
  debug A images
  change parameter order
  skip ref frame
  incorporate illumination map
  apply weights to Y; erode weigth map; remove dark lines for noisy depth
  remove more warnings
  remove unused variables and warnings
  rebase on cmake-restructure
  fix reading in frame list
  read compute_depth_range
  check registration status
  no robust function on prior
  clean up for cr
  add robust_synth.cfg
  move product out of loop
  use srp.ref_frame
  cleanup comments
  replace vil_transform
  duplicate gradient-term-only results
  debug gradient-term-only
  use one or N data terms
  on reference image
  robust/across matrix operation
  merge tv methods in robust branch
  read additional parameters
  add primal-dual steps
  add super_res_robust
  use krtd cameras in exhaustive_depth
  use krtd cameras in super_res_video
  read krtd cameras

+++ super3d/depth/world_frustum.cxx
@@ -81,7 +81,7 @@ world_frustum::warp_cams(const vcl_vector<vpgl_perspective_camera<double> > &cam
   for (unsigned int i = 0; i < newcams.size(); i++)
   {
     newcams[i] = cameras[i];
-    if (i == ref_frame)
+    if (i == (unsigned int) ref_frame)
     {
       vgl_rotation_3d<double> identity;
       identity.set_identity();
